+++ app/controller/loginController.js
@@ -2,7 +2,7 @@ module.exports = class loginController extends nodefony.controller {
 
   constructor(container, context) {
     super(container, context);
-    this.firewall = this.get("security");
+    this.security = this.get("security");
     this.entity = this.getEntity("user");
     this.startSession();
   }
@@ -17,15 +17,18 @@ module.exports = class loginController extends nodefony.controller {
    *    )
    */
   loginAction(type) {
-    let area = this.firewall.getSecuredArea(type);
+    let area = this.security.getSecuredArea(type);
     let action = "/" + type;
     if (area && area.checkLogin) {
       action = area.checkLogin;
     }
+    let google = this.security.getSecuredArea("google_area") ? true : false;
+    let github = this.security.getSecuredArea("github_area") ? true : false;
     return this.render("app:login:login.html.twig", {
       type: type,
       action: action,
-      year: new Date().getFullYear()
+      google: google,
+      github: github
     });
   }
 
@@ -63,7 +66,7 @@ module.exports = class loginController extends nodefony.controller {
    *    @Route ("/logout", name="logout")
    */
   logoutAction() {
-    return this.firewall.logout(this.context)
+    return this.security.logout(this.context)
       .catch((e) => {
         throw e;
       });